refactor(code): create new file for actions and filters
