+++ give-google-analytics.php
@@ -89,6 +89,8 @@ if ( ! class_exists( 'Give_Google_Analytics' ) ) {
 		private function includes() {
 			require_once GIVE_GOOGLE_ANALYTICS_DIR . 'includes/class-give-google-analytics-settings.php';
 			require_once GIVE_GOOGLE_ANALYTICS_DIR . 'includes/give-google-analytics-functions.php';
+			require_once GIVE_GOOGLE_ANALYTICS_DIR . 'includes/filters.php';
+			require_once GIVE_GOOGLE_ANALYTICS_DIR . 'includes/actions.php';
 			require_once GIVE_GOOGLE_ANALYTICS_DIR . 'includes/give-google-analytics-activation.php';
 		}
